Change in EDA and created 4 models
We found that it will be better to work on 3 classes in our target variable. I realised that we didn't learn working on multiple classification problems.

I decided to make two classes in our target variable: Good and bad. Now we don't have a data imbalance problem. 
I was not able to make the model as it was giving me an error. The details in the slack.https://ubc-mds.slack.com/archives/C24J4AVFZ/p1606468720437500
Finally, I made changes in our target variable as True/False. I made a 4 models without any problem. 
The highest accuracy is coming from Random Forest.